refactor use generic class names
