+++ app/src/main/java/com/parishod/watomatic/model/logs/MessageLogsDB.java
@@ -1,0 +1,27 @@
+package com.parishod.watomatic.model.logs;
+
+import android.content.Context;
+
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
+
+import com.parishod.watomatic.model.utils.Constants;
+
+@Database(entities = {MessageLog.class}, version = 1)
+public abstract class MessageLogsDB extends RoomDatabase {
+    private static final String DB_NAME = Constants.WHATSAPP_LOGS_DB_NAME;
+    private static MessageLogsDB _instance;
+
+    public static synchronized MessageLogsDB getInstance(Context context){
+        if(_instance == null){
+            _instance = Room.databaseBuilder(context.getApplicationContext(), MessageLogsDB.class, DB_NAME)
+            .fallbackToDestructiveMigration()
+            .allowMainThreadQueries()
+            .build();
+        }
+        return _instance;
+    }
+
+    public abstract MessageLogsDao logsDao();
+}
