+++ app/src/main/java/com/parishod/watomatic/model/logs/MessageLogsDao.java
@@ -1,14 +1,14 @@
-package com.parishod.watomatic.model.logs.whatsapp;
+package com.parishod.watomatic.model.logs;
 
 import androidx.room.Dao;
 import androidx.room.Insert;
 import androidx.room.Query;
 
 @Dao
-public interface WhatsappAutoReplyLogsDao {
+public interface MessageLogsDao {
     @Query("SELECT created_at FROM WhatsappAutoReplyLogs WHERE user_id=:userId ORDER BY created_at DESC LIMIT 1")
     long getLastReplyTimeStamp(String userId);
 
     @Insert
-    void logReply(WhatsappAutoReplyLogs log);
+    void logReply(MessageLog log);
 }
